Design fixes and Refactor for Quotient page
