+++ components/org/quotient/Header.js
@@ -1,12 +1,20 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faChevronLeft } from '@fortawesome/free-solid-svg-icons';
 import Image from 'next/image';
+import Link from 'next/link';
 
 const Header = () => <div className="p-4 bg-blue-900 flex justify-center items-center">
+    <div className='absolute left-5 cursor-pointer'>
+        <Link href={"/"}>
+            <FontAwesomeIcon className='text-white text-2xl' icon={faChevronLeft} />
+        </Link>
+    </div>
     <Image
         className=''
         src='/quotient.png'
         alt='quotient-logo'
-        width={140}
-        height={30}
+        width={169}
+        height={40}
     />
 </div>;
